mreceive: join IGMP group by interface
mreceive uses the old-style struct ip_mreq for IP_ADD_MEMBERSHIP, which
takes the source address of the interface wishing to join.

Since the IPV6_ADD_MEMBERSHIP variant only takes a struct ipv6_mreq
which contains the ifindex and not the source address, we need to add
support for that.

In preparation for IPv6 support, add logic to join an IGMP group either
by source address or by interface name, whichever is specified.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>