Makefile: make datadir editable
Currently, the datadir is hardcoded to share/doc/mtools. However, multiple
tools named "mtools" exist (e.g., GNU mtools), which can cause conflicts.
When packaging this version of mtools, we have to rename it
(e.g., to mcast-tools) to avoid clashes.

Since the datadir is hardcoded, it can't be adjusted during packaging,
leading to documentation field validation failures. This patch makes the
datadir editable, allowing packagers to modify it as needed.

Signed-off-by: Hangbin Liu <liuhangbin@gmail.com>