Further refactor and code simplification
Move and rename all optional arguments to common.c, localize variables
to their own contexts, check boundaries for arrays and strings, use
variable names that have meaning, drop remaining hungarian notation.

Simplify passing of numbers between msend and receive, encode it as a
string in the message instead of trying to invent broken marshalling.

Finally, drop the timer_cb() in msend in favor of sigwaitinfo(), and
make sure to use the same backend function for sending in all cases.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>